api/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
gunicorn==21.2.0
pymongo==4.4.1
pyjwt==2.8.0
